requirements.txt
numpy==1.26.4
timm
ftfy
regex
iopath
typing_extensions
matplotlib
jupyter
notebook
ipywidgets
ipycanvas
ipympl
pycocotools
decord
opencv-python
einops
scikit-image
scikit-learn
roma
gradio
tqdm
scipy
trimesh
tensorboard
pyglet
huggingface-hub 
pillow
plyfile
safetensors
loguru
monai 
pyrender
smplx 
jpeg4py
pandas
configargparse
flatten_dict
yacs
networkx 
rerun-sdk
ipdb
open3d